### 2018-01-16 #536 @whedon
[PRE REVIEW]: genieR: An R package for inference of demographic history of phylogenies
**Submitting author:** @xiangfstats (<a href="http://orcid.org/0000-0003-3644-7114">Fei Xiang</a>)
**Repository:** <a href="https://github.com/xiangfstats/GenieR" target ="_blank">https://github.com/xiangfstats/GenieR</a>
**Version:** v1.0.0
