### 2018-03-18 #634 @whedon
[REVIEW]: genieR: An R package for inference of demographic history of phylogenies
**Submitting author:** @xiangfstats (<a href="http://orcid.org/0000-0003-3644-7114">Fei Xiang</a>)
**Repository:** <a href="https://github.com/xiangfstats/GenieR" target ="_blank">https://github.com/xiangfstats/GenieR</a>
**Version:** v1.0.0

  ⚠️ **TODO:** No github release exists. DESCRIPTION file lists version as 0.1.0. Please harmonize version numbering and create github release.

Installation works as described on the reviewers machine (Windows 10, R 3.4.2, Rtools 3.5)
⚠️ **Suggestion:** It might be worth pointing out (for the benefit of Windows users - not sure what the target audience is likely using) that a C++ compiler / Rtools installation is necessary. See PR.

⚠️ **Suggestion:** Automated tests are provided covering input checking. All of these appear to be 'negative tests' i.e. based on `expect_error()`. I'd suggest that the authors also implement tests that check the code provides correct output when provided with correct inputs.
