### 2018-05-04 xiangfstats/GenieR#1 @pboesu
[JOSS review] Improvements to documentation
Documentation in this package is quite sparse, and reflects the idea that users have a detailed understanding of the models underlying the GENIE software.  I suggest the following improvements to expand the documentation:
-  [x] A package-level help file (`?genieR` or `?genieR-package`) that gives an overall overview, links/references to the original GENIE software and relevant papers, as well as a link to the package vignette.
-  [x] Function-level help files should have a more detailed explanation of `@param` and  `@value` fields, in particular describing the expected input and provided data structures, e.g. giving relevant S3/S4 classes for phylogeny inputs,  documenting what the components of returned values are (e.g. 'A `data.frame` with columns `t` containing event times and `A` giving lineages [...] for function `att()`).
-  [x] Function explanation for `Geniefit` should be a little more detailed, in particular, providing greater explanations/relevant literature references for the various models that can be fitted.
- [x] Arguments `upper` and `lower` in function `Geniefit` should receive separate `@param` entries in the underlying roxygen2 code. https://github.com/xiangfstats/GenieR/blob/008d3a55cedd1dab6670b1100bbe220dd4d5271f/R/Geniefit.R#L5
- [x] Function description for `att` uses sightly confusing language. The term 'sort out' is colloquial, maybe use 'extract sorted times ...' instead?

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/634
