### 2018-05-04 xiangfstats/GenieR#2 @pboesu
[JOSS review] Improvements to vignette
- [x] The vignette can not be built successfully on my system. The call to `Geniefit` on line 54 (https://github.com/xiangfstats/GenieR/blame/master/vignettes/my-vignette.Rmd#L54) fails with the following error message/traceback
```
Error in if (delfb > -armtst/n) { : missing value where TRUE/FALSE needed
     2. nmkb(start, negllc_log, t = x$t, A = x$A, upper = upper, lower = lower) at Geniefit.R#140
     1. Geniefit(village, Model = "log", start = c(100, 0.1, 0.1), upper = Inf, lower = 0, Rcpp = T)
```
 - [x] Please rename the vignette (currently the default `my-vignette`) to reflect the package name/vignette content.
 - Other useful additions to the vignette would be
    - [x] An explicit benchmarking comparison between model fits obtained using Rcpp vs. R
    - [x] More justification/explanation of the values provided for the `start`, `upper` and `lower` parameters in the `Geniefit` function calls.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/634
