### 2018-05-04 xiangfstats/GenieR#3 @pboesu
[JOSS review] Improvements to API
Some suggenstions to streamline the package API:

- [x] Harmonize function naming conventions. Most function names use snake case, some use dots as separators (`branching.sampling.times`, `heterochronous.gp.stat`), some use camel case or no separator (`MCMCupdate`, `Geniefit`). I would encourage the use of a single naming style.
- [x] Perhaps this comment is simply a reflection of my poor understanding of the underlying phylogenetic methods, but aren't the simulation functions supposed to provide a similar output to that created with the `att()` function from an input phylogeny? If that is the case, it would be nice if `att()` and `coalgen_...` would provide outputs with identical data structures.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/634
