### 2018-05-04 xiangfstats/GenieR#4 @pboesu
[JOSS review] Github release
- [x] Please create a github release of the revised package to facilitate the distribution of a defined version of the package.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/634
