### 2018-05-04 xiangfstats/GenieR#5 @pboesu
[JOSS Review] Improvements to readme
- [x] It might be worth pointing out (for the benefit of Windows users - not sure what the target audience is likely using) that a C++ compiler / Rtools installation is necessary to build this package.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/634
