### 2018-05-04 xiangfstats/GenieR#6 @pboesu
[JOSS review] Improvement to automated tests
Automated tests are provided covering input checking. All of these appear to be 'negative tests' i.e. based on expect_error().
- [x] I'd suggest that the authors also implement tests that check the code provides correct output when provided with correct inputs.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/634
