### 2018-05-04 xiangfstats/GenieR#7 @pboesu
[JOSS review] Improvements to paper
The paper is very terse in it's current form, which is fine in my opinion, as long as the vignette and the documentation are expanded as suggested.

- [x] Please add DOIs wherever possible to the references.

This issue is part of JOSS review https://github.com/openjournals/joss-reviews/issues/634
