### 2018-05-04 #634 @pboesu
Hello @biorelated and @xiangfstats

I have reviewed the package now and provided a number of issues in the target repositories that, in my opinion, warrant responses from the authors.

Critically, the vignette does not compile because of an error in one of the package functions (https://github.com/xiangfstats/GenieR/issues/2), but I would also suggest a number of fairly minor changes to the documentation, API, and testing suite that should greatly improve the re-use potential of the package.
