### 2018-05-05 #634 @biorelated
Thank you very much @pboesu  for the review.
 @xiangfstats kindly respond to the raised issues.
