### 2018-06-13 d6922b2
1st review from JOSS
