+++ R/genieR-package.R
@@ -1,0 +1,43 @@
+#' genieR: Inference of demographic history from phylogeny
+#'
+#' @description The R package genieR allows inference of demographic history from reconstructed molecular phylogenies.
+#' The package provides a statistical inference based on coalescent theory for the demographic history of a population from
+#' phyloginies that have been reconstructed from sampled DNA sequences. It covers a few functions designed in
+#' GENIE, Pybus and Rambaut (2002).
+#'
+#' The inference is conducted in both the maximum likelihood approach and Bayesian approach. The package also allow
+#' users to simulates coalescen and sampling times for isochronous and heterochronous data. Furthermore, users can extract
+#' sorted sampling/coalescent times and sampling/coalescent lineages from a phylogenetic tree.
+#'
+#' We encourage users to look at \code{\link{genieR.Rmd}}.
+#'
+#' @docType package
+#'
+#' @name genieR
+#'
+#'
+#'
+#'
+#'
+#' @references
+#'
+#' Hastings, W. K. (1970). Monte Carlo Sampling Methods Using Markov Chains and Their Applications.Biometrika 57 (1): 97-109.
+#'
+#' Kingman, J. F. C. (1982). On the Genealogy of Large Populations. Journal of Applied Probability 19, 27-43.
+#'
+#' Griffiths, R., and Tavare, S. (1994). Sampling Theory for Neutral Alleles in a Varying Environment. Philosophical Transactions: Biological Sciences, 344(1310), 403-410.
+#'
+#' Pybus, O. G., and Rambaut, A. (2002). GENIE: Estimating Demographic History from Molecular Phylogenies. Bioinformatics 18, 1404-1405.
+#'
+#' \url{https://github.com/JuliaPalacios/coalsieve}.
+#'
+#'
+#' @author Fei Xiang, Bethany Dearlove and Simon D. W. Frost.
+#'
+#'
+#'       Maintainer: Fei Xiang (\email{xf3087@@gmail.com}),
+#'
+#'
+
+
+   NULL
