### 2018-06-13 3ed9080
1st review from JOSS
