+++ man/coalgen.iso.Rd
@@ -1,0 +1,34 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/coalgen_iso.R
+\name{coalgen.iso}
+\alias{coalgen.iso}
+\title{Simulation of  coalescent times for isochronous data.}
+\usage{
+coalgen.iso(sample, trajectory, val_upper = 10)
+}
+\arguments{
+\item{sample}{A two dimensional vector of number of individuals and the initial time.}
+
+\item{trajectory}{A population growth function.}
+
+\item{val_upper}{Upper end of time points to be simulated.}
+}
+\value{
+A data.frame with columns t containing event times and A giving lineages.
+}
+\description{
+\code{coalgen.iso} simulates coalescent times for isochronous data.
+}
+\examples{
+sample<-c(100,0)
+
+trajectory<-function(x)  exp(10*x)
+example.iso<-coalgen.iso(sample, trajectory)
+
+}
+\references{
+\url{https://github.com/JuliaPalacios/coalsieve}.
+}
+\author{
+Fei Xiang (\email{xf3087@gmail.com})
+}
