+++ man/coalgen.hetero.Rd
@@ -1,0 +1,33 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/coalgen_hetero.R
+\name{coalgen.hetero}
+\alias{coalgen.hetero}
+\title{Simulation of coalescent times for heterochronous data.}
+\usage{
+coalgen.hetero(sample, trajectory, val_upper = 10)
+}
+\arguments{
+\item{sample}{A two columns matrix of number of individuals and the initial time.}
+
+\item{trajectory}{A population growth function.}
+
+\item{val_upper}{Upper end of time points to be simulated.}
+}
+\value{
+A data.frame with columns t containing event times and A giving lineages.
+}
+\description{
+\code{coalgen.hetero} simulates coalescent times for heterochronous data.
+}
+\examples{
+sample1<-cbind(c(9,1,2,1),c(0,.008,.03,.1))
+
+trajectory<-function(x)  exp(10*x)
+example.hetero<-coalgen.hetero(sample1, trajectory)
+}
+\references{
+\url{https://github.com/JuliaPalacios/coalsieve}.
+}
+\author{
+Fei Xiang (\email{xf3087@gmail.com})
+}
