### 2018-06-13 23f00fd
Delete my-vignette.Rmd
