+++ vignettes/my-vignette.Rmd
@@ -1,76 +1,0 @@
----
-title: "Introduction to genieR"
-author: "Fei Xiang"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Vignette Title}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-This vignette demonstrates the use of the package. An HIV phylogenetic tree is included in the package directory. To visualise the tree:
-
-```{r message=FALSE}
-library(genieR)
-```
-
-```{r}
-library(ape)
-data(village)
-plot(village,show.tip.label = F,edge.color = sample(colors(), length(village$edge)/2,replace=T))
-title("A phylogenetic tree with random colors")
-```
-
-The package can sort out sampling and coalescent times by 
-```{r}
-x=att(village)
-head(x)
-tail(x)
-```
-The first column of x is the times and the second column of x is the count of current samples, which increases by one representing a sampling events and decreases by one representing a coalescent event. To see the times of coalescents,
-```{r}
-x$t[diff(x[,2])==-1]
-```
-and the times of sampling,
-```{r}
-x$t[diff(x[,2])==1]
-```
-
-A simpler way of extract sampling and coalescents times is through function 
-```{r}
-heterochronous.gp.stat(village)
-```
-
-
-Apart from collecting the basic information of a phylogenetic tree, we can fit a exponential growth coalescent model for this tree by Genefit, 
-```{r}
-fit1=Geniefit(village,Model="expo",start=c(100,.1),upper=Inf,lower=0)
-fit1
-```
-The output includes estimated parameters, corresponding logliklihood and the AIC information of this exponetial growth model. Other coalescent models to choose are  const (constant population size),expan (expansion growth), log (logistic growth), step (piecewise constant), pexpan (piecewise expansion growth) and plog (piecewise logistic growth). Genifit allows users to infer parameters through c++ code, for example,
-```{r}
-library(dfoptim)
-fit2=Geniefit(village,Model="log",start=c(100,.1,.1),upper=Inf,lower=0,Rcpp=T)
-fit2
-```
-The AIC suggests the logistic model is a better fit. The function can also produce a MCMC fit to a coalescent model.
-```{r}
-fit3=Geniefit(village,Model="log",start=c(100,.1,.1),upper=Inf,lower=0,MCMC=T)
-fit3
-```
-Obviousely, The MCMC fit is better as it provides a better loglikelihood under the logistic model.
-
-This package can also simulates coalescent times for isochronous/heterochronous data. 
-```{r}
- trajectory<-function(x)  exp(10*x)
- sample1<-cbind(c(9,1,2,1),c(0,.008,.03,.1))
- example_hetero<-coalgen_hetero(sample1, trajectory)
- example_hetero
- sample<-c(10,0)
- example_iso<-coalgen_iso(sample, trajectory)
- example_iso
-```
-
-
-
