### 2018-06-13 bbb2904
1st review from JOSS
