+++ paper.bib
@@ -18,6 +18,7 @@
   year          = {1982},
   volume        = {19},
   pages         = {27-43},
+  doi           = {10.2307/3213548},
 }
 
 @Article{PybusRambautGENIE,
@@ -27,6 +28,7 @@
   year          = {2002},
   volume        = {18},
   pages         = {1404-1405},
+  doi           = {10.1093/bioinformatics/18.10.1404},
 }
 
 @Article{Hastings1970,
@@ -37,6 +39,7 @@
   volume        = {57},
   number        = {1},
   pages         = {97-109},
+  doi           = {10.2307/2334940},
 }
 
 @Article{Bethanycluster2017,
