+++ NAMESPACE
@@ -1,9 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
-export(Geniefit)
 export(att)
-export(coalgen_hetero)
-export(coalgen_iso)
+export(coalgen.hetero)
+export(coalgen.iso)
+export(geniefit)
 export(heterochronous.gp.stat)
 export(loglik)
 export(negllc_const)
