### 2018-06-13 xiangfstats/GenieR#2 @xiangfstats
Dear @pboesu
Thanks for the comments on the vignette. The renamed  vignette can be built now.  I also added a benchmarking comparison model fits obtained using Rcpp vs. R and more explanation of the values fro the `start`, `upper` and `lower` parameters in the `geniefit` function calls
Best
Fei
