### 2018-06-13 xiangfstats/GenieR#3 @xiangfstats
Dear @pboesu
Thanks for the comments on the API. I rename most functions using dots as separators and rename `Geniefit` to `geniefit`. The output of `att`, `coalgen.iso` and `coalgen.hetero` now provide outputs with same data structure.
Best
Fei
