### 2018-06-13 xiangfstats/GenieR#4 @xiangfstats
Dear @pboesu
Thanks for the comments on Github release. I created a github release.
Best
Fei
