### 2018-06-13 xiangfstats/GenieR#5 @xiangfstats
Dear @pboesu
Thanks for the comments on readme. I added that  Rtools installation is necessary to build this package.
Best
Fei
