### 2018-06-13 xiangfstats/GenieR#6 @xiangfstats
Dear @pboesu
Thanks for the comments on tests. I added some positive tests on the correct output by expect_true()
Best
Fei
