### 2018-06-13 xiangfstats/GenieR#7 @xiangfstats
Dear @pboesu
Thanks for the comments on the paper. I added DOIs for some references.
Best
Fei
