### 2018-06-13 xiangfstats/GenieR#1 @xiangfstats
Dear @pboesu
Thanks for comments on the documentation. I added a package-level help file including an overview, references to GENIE, relevant papers and a link to the package vignette. Function-level help files now have more detailed explanation of `@param` and `@value`.  Function explanation of `geniefit` is more detailed and arguments `upper` and `lower` in `geniefit` are separated. Function description for `att` now uses 'extract sorted times...'
Best
Fei
