### 2018-06-13 #634 @xiangfstats
Dear @george-githinji
I have made improvements based on @pboesu 's suggestions and replied separately to those 7 comments.
Thanks very much for the review from @pboesu .
Best
Fei
