### 2018-07-08 #634 @george-githinji
Dear @pboesu Please review the improvements made by @xiangfstats. Kindly let me know if you need some assistance or if you are not in a position to. Thank you.
