### 2019-01-05 #634 @xiangfstats
Dear @pjotrp ,
I can confirm that all references have a DOI except two without DOIs. I created a Github release and deposit that release with Zenodo. The DOI is  DOI: 10.5281/zenodo.2532305. Many thanks
Best
Fei
