### 2019-01-07 #634 @xiangfstats
Thanks very much to everyone, @arfon, @pboesu, @pjotrp and @george-githinji. Great to know our paper is accepted in JOSS.
